refactor: extract SessionArtifactsManaging+Defaults.swift
Splits the `SessionArtifactsManaging` default-impl extension
(preserveRecordedAudio, ~30 lines) out of SessionArtifactsService.swift
into its own file, matching the +Defaults naming convention used
elsewhere in the codebase.

Byte-identical move of lines 112-140. Both files import Foundation.

SessionArtifactsService.swift: 140 → 110 lines. Tests: 667.
